Moved `ReflectKind` and friends to new module
